+++ dynamic_forms/migrations/0005_auto_increase_form_label_max_length_255.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import models, migrations
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('dynamic_forms', '0004_formmodel_recipient_email'),
+    ]
+
+    operations = [
+        migrations.AlterField(
+            model_name='formfieldmodel',
+            name='label',
+            field=models.CharField(max_length=255, verbose_name='Label'),
+            preserve_default=True,
+        ),
+    ]